requirements.txt
mkdocs-material==9.5.*
mike==2.*
